feat(row): support nested, improve test resilience (#331)
* test(*): review transaction id usage

* fix(api): further transaction id usage review

* fix(tests): mongodb collation setting to accommodate our locales

* test(lexicaleditorwithmultipleblocks): create media for fixture

* feat(plugin): crowdin files delete option

* test(lexicaleditorwithblocks): fix directory expectations

* test(lexicaleditorwithmultipleblocks): media logic and snapshots

* chore(dev): regenerate importmap

* feat(document): restore contribution to update local file record

* chore(dev): regenerate importmap

* fix(document): restore contributor change

* refactor(getcollapsiblelocalizedfields): merge fns

* 📝 CodeRabbit Chat: Add unit tests

* 📝 CodeRabbit Chat: Add unit tests

---------

+++ plugin/src/lib/api/files/index.ts
@@ -160,6 +160,7 @@ export class payloadCrowdinSyncFilesApi {
             `Found existing file on Crowdin. File ID: ${existingFile.data.id}`,
           );
           // Return the existing file so it can be synced to the local database
+          (existingFile as any)._payloadCrowdinSyncWasExisting = true;
           return existingFile;
         } else {
           console.error(
@@ -201,10 +202,12 @@ export class payloadCrowdinSyncFilesApi {
     ) {
       return;
     }
-    await this.sourceFilesApi.deleteDirectory(
-      this.projectId,
-      crowdinPayloadArticleDirectory.originalId,
-    );
+    if (this.pluginOptions.deleteCrowdinFiles) {
+      await this.sourceFilesApi.deleteDirectory(
+        this.projectId,
+        crowdinPayloadArticleDirectory.originalId,
+      );
+    }
     await this.req.payload.delete({
       collection: 'crowdin-article-directories',
       id: crowdinPayloadArticleDirectory.id,
